# Improve printing of several arguments