Following a discussion on [Discourse](https://discourse.julialang.org/t/string-optimisation-in-julia/119301/10?u=gdalle), this PR tries to improve `print` (and variants) for more than one argument.
The idea is that `for` is type-unstable over the tuple `args`, while `foreach` unrolls. I have no idea how to properly evaluate the potential performance gains though.