@Tortar 2024-09-12
To test if it works you could execute https://github.com/shagrouni/langs_string_build_test/blob/main/src/jl_strbld.jl with `julia jl_strbld.jl 50000000` and see if it now works faster or as fast as using 

```julia
    ...
    for i in 1:num
        print(buf, " J ")
        print(buf, i)
    end
    ...
```