@gdalle 2024-09-12
Doesn't look half bad.

```julia
using BenchmarkTools

function test_slow(n::Integer)
    s = "hello"
    buf = IOBuffer()
    for i in 1:n
        print(buf, s, i)
    end
end

function test_fast(n::Integer)
    s = "hello"
    buf = IOBuffer()
    for i in 1:n
        print(buf, s)
        print(buf, i)
    end
end

@btime test_slow(50_000);
@btime test_fast(50_000);
```

**Current nightly**

```julia
julia> @btime test_slow(50_000);
  5.376 ms (249003 allocations: 7.35 MiB)

julia> @btime test_fast(50_000);
  2.959 ms (100025 allocations: 4.31 MiB)
```

**This PR**

```julia
julia> @btime test_slow(50_000);
  3.503 ms (100025 allocations: 4.31 MiB)

julia> @btime test_fast(50_000);
  3.018 ms (100025 allocations: 4.31 MiB)
```