@stevengj 2024-09-12
Ideally, `test_slow` and `test_fast` should now be equivalent.  @tkf, shouldn't `foreach` now be equivalent to a fully unrolled loop thanks to #31901?