@Tortar 2024-09-12
Does the technique has any downside @stevengj such as increasing compile time substantially? Because otherwise this can probably be applied in other `Vararg` cases right?