@nsajko 2024-09-13
The `_str_sizehints` changes sadly break bootstrap for the assert builds.

```
gmp.jl
ryu/Ryu.jl
ryu/utils.jl
ryu/shortest.jl
Internal error: during type inference of
string(Expr)
Encountered unexpected error in runtime:
AssertionError(msg="irinterp is unable to handle heavy recursion")
collect_limitations! at ./compiler/ssair/irinterp.jl:4 [inlined]
from_interprocedural! at ./compiler/abstractinterpretation.jl:392 [inlined]
```