@gdalle 2024-09-13
> The _str_sizehints changes sadly break bootstrap for the assert builds.

@nsajko what does that mean in layperson terms?