@nsajko 2024-09-13
The bootstrap (the initial compilation of `Base` and other code) fails because the compiler refuses to handle deep recursion. Or something like that.

In practice that probably means you have to revert the commit I suggested (I think the build was successful for the initial commit in the PR, right?). I guess we could try reintroducing the `mapreduce` after #55575 gets merged.