@vtjnash 2024-09-13
This adds a dynamic closure capture though and dispatch, which may not be performant