@JeffBezanson 2024-09-13
That error isn't supposed to happen; you may have found an interesting bug :)