@gdalle 2024-09-13
@vtjnash I think it comes down to a compromise between tuple unrolling (which is possible with foreach but not with for) and the additional closure required. My PR might be slower for printing tuples with identically typed arguments for instance. I just wanted to experiment with Steven's suggestion but I don't have strong feeling either way, so I'll let the knowledgeable people hash it out.