@stevengj 2024-09-18
What does the benchmark look like with the latest `Fix1` version?  Is it matching the manually unrolled 2-`print`-call version yet?