@LilithHafner 2024-09-19
Compile time regression is also gone!

```julia-repl
julia> compile_time(version) = parse(Float64, strip(read(`julia +$version -E 'function test_slow(n::Integer)
               s = "hello"
               buf = IOBuffer()
               for i in 1:n
                   print(buf, s, i)
               end
           end; @elapsed @eval test_slow(100)'`, String)))
compile_time (generic function with 1 method)

julia> [compile_time("nightly") for _ in 1:7]
Juliaup configuration is locked by another process, waiting for it to unlock.
Juliaup configuration is locked by another process, waiting for it to unlock.
Juliaup configuration is locked by another process, waiting for it to unlock.
7-element Vector{Float64}:
 0.006195663
 0.005031322
 0.005418033
 0.005152656
 0.005105739
 0.005199073
 0.005210907

julia> [compile_time("pr55754") for _ in 1:7]
7-element Vector{Float64}:
 0.008983305
 0.005014739
 0.004937614
 0.005452533
 0.005186157
 0.005184865
 0.005381825

shell> julia +nightly -E 'Base.GIT_VERSION_INFO.commit'
"58b239c5b8eaffec3b9b99e6d7c37e8ae6129d6d"

shell> julia +pr55754 -E 'Base.GIT_VERSION_INFO.commit'
"91443de63550381d4d95d7b992e3d9de8206e8c3"
```